Find the middle of a given linked list
For given a singly linked list, find the middle of the linked list. 
e.g, if the given linked list is 6->7->8->9->10 then the output should be 8. 
If there are even nodes, then there would be two middle nodes, we need to print the second middle element. For example, if the given linked list is 6->7->8->9->10->11 then the output should be 9.